more commands/inquiries

+++ src/tests/unit/protocol/message/test_binary.cpp
@@ -8,7 +8,7 @@
 
 TEST_CASE("BinaryDataBuffer::BinaryDataBuffer()", "[Protocol][Protocol::Message]")
 {
-  auto size = GENERATE(static_cast<uint32_t>(10), static_cast<uint8_t>(1000));
+  auto size = GENERATE(static_cast<uint64_t>(10), static_cast<uint64_t>(1000));
   {
 
     std::vector<irsol::protocol::internal::byte_t> data(size);
@@ -25,10 +25,10 @@ TEST_CASE("ImageBinaryData::ImageBinaryData()", "[Protocol][Protocol::Message]")
   auto sizeData = GENERATE(
     std::make_pair(
       static_cast<uint64_t>(160),
-      std::array<uint32_t, 2>{static_cast<uint32_t>(10), static_cast<uint32_t>(16)}),
+      std::array<uint64_t, 2>{static_cast<uint64_t>(10), static_cast<uint64_t>(16)}),
     std::make_pair(
       static_cast<uint64_t>(921600),
-      std::array<uint32_t, 2>{static_cast<uint32_t>(1280), static_cast<uint32_t>(720)}));
+      std::array<uint64_t, 2>{static_cast<uint64_t>(1280), static_cast<uint64_t>(720)}));
   {
     std::vector<irsol::protocol::internal::byte_t> data(sizeData.first);
     irsol::protocol::ImageBinaryData               m{std::move(data), sizeData.second};
@@ -44,12 +44,12 @@ TEST_CASE("ColorImageBinaryData::ColorImageBinaryData()", "[Protocol][Protocol::
   auto sizeData = GENERATE(
     std::make_pair(
       static_cast<uint64_t>(160),
-      std::array<uint32_t, 3>{
-        static_cast<uint32_t>(10), static_cast<uint32_t>(16), static_cast<uint32_t>(1)}),
+      std::array<uint64_t, 3>{
+        static_cast<uint64_t>(10), static_cast<uint64_t>(16), static_cast<uint64_t>(1)}),
     std::make_pair(
       static_cast<uint64_t>(2764800),
-      std::array<uint32_t, 3>{
-        static_cast<uint32_t>(1280), static_cast<uint32_t>(720), static_cast<uint32_t>(3)}));
+      std::array<uint64_t, 3>{
+        static_cast<uint64_t>(1280), static_cast<uint64_t>(720), static_cast<uint64_t>(3)}));
   {
     std::vector<irsol::protocol::internal::byte_t> data(sizeData.first);
     irsol::protocol::ColorImageBinaryData          m{std::move(data), sizeData.second};
